requirements.txt
metatrader5==5.0.5509
numpy==2.4.1
django>=4.2
djangorestframework>=3.14
drf-spectacular>=0.27
